[db,parsers] move read_fonts_conf to parsers
